+++ packages/synthchron-tauri/src/components/NewProjectModal.tsx
@@ -1,35 +1,9 @@
 import { useState } from 'react'
 
-import { faker } from '@faker-js/faker'
-import {
-  Box,
-  Button,
-  FormControl,
-  Input,
-  InputLabel,
-  MenuItem,
-  Modal,
-  Select,
-  SelectChangeEvent,
-  Stack,
-  TextField,
-  Typography,
-} from '@mui/material'
-import { useNavigate } from 'react-router-dom'
+import { Box, Modal, Tab, Tabs } from '@mui/material'
 
-import {
-  PetriNetProcessModel,
-  ProcessModel,
-  ProcessModelType,
-} from '@synthchron/simulator'
-
-import { usePersistentStore } from './common/persistentStore'
-
-export type ProjectConfig = {
-  name: string
-  description: string
-  modelType: ProcessModelType
-}
+import NewProjectDefault from './NewProjectDefault'
+import NewProjectFile from './NewProjectFile'
 
 const modal_style = {
   position: 'absolute' as const,
@@ -43,64 +17,14 @@ const modal_style = {
   p: 4,
 }
 
+const tab_style = {
+  fontSize: '0.8rem',
+}
+
 interface NewProjectModalProps {
   open: boolean
   onClose: () => void
   redirect?: boolean
-}
-
-const examplePetriNetModel: PetriNetProcessModel = {
-  type: ProcessModelType.PetriNet,
-  acceptingExpressions: [
-    {
-      name: 'accept',
-      expression: 'p2 >= 7',
-    },
-  ],
-  nodes: [
-    {
-      type: 'place',
-      name: 'p1',
-      identifier: 'p1',
-      amountOfTokens: 5,
-      position: {
-        x: -200,
-        y: 0,
-      },
-    },
-    {
-      type: 'transition',
-      name: 'transition',
-      identifier: 't1',
-      weight: 1,
-      position: {
-        x: 0,
-        y: 0,
-      },
-    },
-    {
-      type: 'place',
-      name: 'p2',
-      identifier: 'p2',
-      amountOfTokens: 0,
-      position: {
-        x: 200,
-        y: 0,
-      },
-    },
-  ],
-  edges: [
-    {
-      source: 'p1',
-      target: 't1',
-      multiplicity: 1,
-    },
-    {
-      source: 't1',
-      target: 'p2',
-      multiplicity: 2,
-    },
-  ],
 }
 
 const NewProjectModal: React.FC<NewProjectModalProps> = ({
@@ -108,77 +32,7 @@ const NewProjectModal: React.FC<NewProjectModalProps> = ({
   onClose,
   redirect = false,
 }) => {
-  const navigate = useNavigate()
-
-  const [processModel, setProcessModel] = useState({ type: 'none' })
-
-  const addProject = usePersistentStore((state) => state.addProject)
-
-  const newProjectDefault = () => {
-    return {
-      name: faker.animal.bird(),
-      description: '',
-      modelType: ProcessModelType.PetriNet,
-    }
-  }
-
-  const [newProjectConfig, setNewProjectConfig] = useState<ProjectConfig>(
-    newProjectDefault()
-  )
-
-  const updateNewProjectConfig = (fields: Partial<ProjectConfig>) => {
-    setNewProjectConfig({ ...newProjectConfig, ...fields })
-  }
-
-  const createNewProject = () => {
-    let model: ProcessModel
-    switch (newProjectConfig.modelType) {
-      default:
-      case ProcessModelType.PetriNet:
-        model = examplePetriNetModel
-        break
-      case ProcessModelType.DcrGraph:
-        model = { type: newProjectConfig.modelType, nodes: [], edges: [] }
-        break
-      case ProcessModelType.Flowchart:
-        model = {
-          type: newProjectConfig.modelType,
-          nodes: [
-            {
-              type: 'decision',
-              identifier: 'startNode',
-            },
-          ],
-          edges: [],
-          initialNode: 'startNode',
-        }
-        break
-    }
-
-    if (processModel.type != null) {
-      model = processModel as ProcessModel
-    }
-
-    const projectId = addProject({
-      projectName: newProjectConfig.name,
-      projectDescription: newProjectConfig.description,
-      projectModel: model, // TODO
-      created: new Date().toJSON(),
-      lastEdited: new Date().toJSON(),
-      lastOpened: (redirect ? new Date() : new Date(0)).toJSON(),
-    })
-    setNewProjectConfig(newProjectDefault())
-    if (redirect) navigate(`/editor/${projectId}`)
-    onClose()
-  }
-
-  const readFile = (file: File) => {
-    console.log(file)
-    file.text().then((res: string) => {
-      const json: ProcessModel = JSON.parse(res) as ProcessModel
-      setProcessModel(json)
-    })
-  }
+  const [value, setValue] = useState(0)
 
   return (
     <Modal
@@ -187,75 +41,15 @@ const NewProjectModal: React.FC<NewProjectModalProps> = ({
       aria-labelledby='modal-modal-title'
       aria-describedby='modal-modal-description'
     >
-      <Box
-        component='form'
-        sx={modal_style}
-        noValidate
-        onSubmit={(event: React.FormEvent<HTMLFormElement>) => {
-          event.preventDefault()
-          createNewProject()
-        }}
-      >
-        <Typography
-          id='modal-modal-title'
-          variant='h5'
-          sx={{ marginBottom: '2rem' }}
-        >
-          New project
-        </Typography>
-        <Stack spacing={4} direction='column'>
-          <TextField
-            required
-            id='new-project-name'
-            label='Project name'
-            variant='standard'
-            value={newProjectConfig.name}
-            onChange={(event: React.ChangeEvent<HTMLInputElement>) => {
-              updateNewProjectConfig({ name: event.target.value })
-            }}
-          />
-          <Input
-            type='file'
-            className='Mui-error'
-            onChange={(e) => {
-              if (e != null && e.target != null) {
-                if (e.target.files != null && e.target.files.length > 0)
-                  readFile(e.target.files[0])
-              }
-            }}
-          ></Input>
-          <FormControl fullWidth>
-            <InputLabel id='demo-simple-select-label'>Model Type</InputLabel>
-            <Select
-              labelId='demo-simple-select-label'
-              id='new-project-model'
-              label='Process Model Type'
-              value={newProjectConfig.modelType}
-              onChange={(event: SelectChangeEvent) => {
-                updateNewProjectConfig({
-                  modelType: event.target.value as ProcessModelType,
-                })
-              }}
-            >
-              <MenuItem value={ProcessModelType.PetriNet}>Petri Net</MenuItem>
-              <MenuItem value={ProcessModelType.Flowchart}>Flowchart</MenuItem>
-              <MenuItem value={ProcessModelType.DcrGraph}>Dcr Graph</MenuItem>
-            </Select>
-          </FormControl>
-
-          <TextField
-            id='new-project-description'
-            label='Description'
-            multiline
-            value={newProjectConfig.description}
-            onChange={(event: React.ChangeEvent<HTMLInputElement>) => {
-              updateNewProjectConfig({ description: event.target.value })
-            }}
-          />
-          <Button type='submit' color='primary' variant='contained'>
-            Create
-          </Button>
-        </Stack>
+      <Box sx={modal_style}>
+        <Tabs value={value} onChange={(_event, value) => setValue(value)}>
+          <Tab sx={tab_style} label={'Create empty project'} />
+          <Tab sx={tab_style} label={'Create project from file'} />
+        </Tabs>
+        {value == 0 && (
+          <NewProjectDefault onClose={onClose} redirect={redirect} />
+        )}
+        {value == 1 && <NewProjectFile onClose={onClose} redirect={redirect} />}
       </Box>
     </Modal>
   )
